+++ src/ext/lib/Logger.js
@@ -227,29 +227,21 @@ class Logger {
     stackInfo['mousemove'] = false;
     stackInfo['exitLogs'] = false;
 
-    // here we check which source triggered the action. We know that only one of these
-    // functions will appear in the trace at most once (and if more than one of these
-    // appears — e.g. frameCheck triggered by user toggling autodetection in popup —
-    // the most recent one will be the correct one 99% of the time)
+    // here we check which source triggered the action. There can be more
+    // than one source, too, so we don't break when we find the first one
     for (const line of stackInfo.stack.trace) {
       if (line === 'doPeriodicPlayerElementChangeCheck') {
         stackInfo['periodicPlayerCheck'] = true;
-        break;
       } else if (line === 'doPeriodicFallbackChangeDetectionCheck') {
         stackInfo['periodicVideoStyleChangeCheck'] = true;
-        break;
       } else if (line === 'frameCheck') {
         stackInfo['aard'] = true;
-        break;
       } else if (line === 'execAction') {
         stackInfo['keyboard'] = true;
-        break;
       } else if (line === 'processReceivedMessage') {
         stackInfo['popup'] = true;
-        break;
       } else if (line === 'handleMouseMove') {
         stackInfo['mousemove'] = true;
-        break;
       }
     }
 
@@ -374,9 +366,21 @@ class Logger {
     });
   }
 
-  logToConsole(message, stackInfo) {
+  logToConsole(level, message, stackInfo) {
     try {
-      console.log(...message, {stack: stackInfo});
+      switch (level) {
+        case 'error': 
+          console.error(...message, {stack: stackInfo});
+          break;
+        case 'warn':
+          console.warn(...message, {stack: stackInfo});
+          break;
+        case 'info': 
+          console.info(...message, {stack: stackInfo});
+          break;
+        default: 
+          console.log(...message, {stack: stackInfo});
+      }
     } catch (e) {
       console.error("Message too big to log. Error:", e, "stackinfo:", stackInfo);
     }
@@ -399,7 +403,7 @@ class Logger {
         this.logToFile(message, stackInfo);
       }
       if (this.conf.consoleOptions?.enabled) {
-        this.logToConsole(message, stackInfo);
+        this.logToConsole(level, message, stackInfo);
       }
       return; // don't check further — recursion-land ahead!
     }
@@ -420,7 +424,7 @@ class Logger {
     }
     if (this.conf.consoleOptions?.enabled) {
       if (this.canLogConsole(component) || stackInfo.exitLogs) {
-        this.logToConsole(message, stackInfo);
+        this.logToConsole(level, message, stackInfo);
       }
     }
   }