+++ CHANGELOG.md
@@ -13,7 +13,12 @@ QoL improvements for me:
 
 * logging: allow to enable logging at will and export said logs to a file
 
-### v4.4.8 (Current)
+### v4.4.9 (Current)
+
+* Fixed the youtube alignment issue (previously fixed in v4.4.7.1-2), but this time for real (and in a bit more proper way)
+* Fixed the bug where extension wouldn't work when URL specified a port (e.g. www.example.com:80)
+
+### v4.4.8
 
 * Fixed the bug where on pages with more than one video, the list of available videos in the extension popup wouldn't remove videos that are no longer displayed on site. This resulted in extension listing videos that were no longer on the page. Reboot or navigation would also not clear the list if navigating between various pages on the same host.
 * Fixed the chrome-only bug where on sites with more than one video, the number wouldn't get hidden when the extension popup closed.